Extract HTTP server file upload tests in a single class
