+++ src/test/java/io/vertx/core/http/Http1xServerFileUploadTest.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2011-2019 Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ */
+package io.vertx.core.http;
+
+import org.junit.Ignore;
+import org.junit.Test;
+
+/**
+ */
+public class Http1xServerFileUploadTest extends HttpServerFileUploadTest {
+
+
+  @Ignore
+  @Test
+  @Override
+  public void testBrokenFormUploadLargeFile() {
+    super.testBrokenFormUploadLargeFile();
+  }
+
+  @Ignore
+  @Test
+  @Override
+  public void testBrokenFormUploadLargeFileStreamToDisk() {
+    super.testBrokenFormUploadLargeFileStreamToDisk();
+  }
+
+}
