+++ src/test/java/io/vertx/core/http/Http2ServerFileUploadTest.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2011-2019 Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ */
+package io.vertx.core.http;
+
+/**
+ */
+public class Http2ServerFileUploadTest extends HttpServerFileUploadTest {
+
+  @Override
+  protected HttpServerOptions createBaseServerOptions() {
+    return Http2TestBase.createHttp2ServerOptions(DEFAULT_HTTP_PORT, DEFAULT_HTTP_HOST);
+  }
+
+  @Override
+  protected HttpClientOptions createBaseClientOptions() {
+    return Http2TestBase.createHttp2ClientOptions();
+  }
+
+}
