+++ src/test/java/io/vertx/core/http/HttpTest.java
@@ -2996,323 +2996,6 @@ public abstract class HttpTest extends HttpTestBase {
   }
 
   @Test
-  public void testFormUploadEmptyFile() {
-    testFormUploadFile("", false, false);
-  }
-
-  @Test
-  public void testFormUploadSmallFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(100), false, false);
-  }
-
-  @Test
-  public void testFormUploadMediumFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(20000), false, false);
-  }
-
-  @Test
-  public void testFormUploadLargeFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(4 * 1024 * 1024), false, false);
-  }
-
-  @Test
-  public void testFormUploadEmptyFileStreamToDisk() {
-    testFormUploadFile("", true, false);
-  }
-
-  @Test
-  public void testFormUploadSmallFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(100), true, false);
-  }
-
-  @Test
-  public void testFormUploadMediumFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(20 * 1024), true, false);
-  }
-
-  @Test
-  public void testFormUploadLargeFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(4 * 1024 * 1024), true, false);
-  }
-
-  @Test
-  public void testFormUploadWithExtFilename() {
-    testFormUploadFile(null, "%c2%a3%20and%20%e2%82%ac%20rates", "the-content", true, false);
-  }
-
-  @Test
-  public void testBrokenFormUploadEmptyFile() {
-    testFormUploadFile("", true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadSmallFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(100), true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadMediumFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(20 * 1024), true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadLargeFile() {
-    testFormUploadFile(TestUtils.randomAlphaString(4 * 1024 * 1024), true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadEmptyFileStreamToDisk() {
-    testFormUploadFile("", true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadSmallFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(100), true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadMediumFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(20 * 1024), true, true);
-  }
-
-  @Test
-  public void testBrokenFormUploadLargeFileStreamToDisk() {
-    testFormUploadFile(TestUtils.randomAlphaString(4 * 1024 * 1024), true, true);
-  }
-
-  private void testFormUploadFile(String contentStr, boolean streamToDisk, boolean abortClient) {
-    testFormUploadFile("tmp-0.txt", "tmp-0.txt", contentStr, streamToDisk, abortClient);
-  }
-
-  private void testFormUploadFile(String filename,
-                                  String extFilename,
-                                  String contentStr,
-                                  boolean streamToDisk,
-                                  boolean abortClient) {
-    String expectedFilename;
-    try {
-      if (extFilename != null) {
-        expectedFilename = URLDecoder.decode(extFilename, "UTF-8");
-      } else {
-        expectedFilename = filename;
-      }
-    } catch (UnsupportedEncodingException e) {
-      fail(e);
-      return;
-    }
-
-    waitFor(2);
-
-    Buffer content = Buffer.buffer(contentStr);
-
-    AtomicInteger attributeCount = new AtomicInteger();
-
-    AtomicReference<HttpConnection> clientConn = new AtomicReference<>();
-    AtomicReference<HttpConnection> serverConn = new AtomicReference<>();
-    Runnable checkClose = () -> {
-      if (clientConn.get() != null && serverConn.get() != null) {
-        clientConn.get().close();
-      }
-    };
-
-    server.requestHandler(req -> {
-      if (req.method() == HttpMethod.POST) {
-        assertEquals(req.path(), "/form");
-        req.response().setChunked(true);
-        req.setExpectMultipart(true);
-        assertTrue(req.isExpectMultipart());
-
-        // Now try setting again, it shouldn't have an effect
-        req.setExpectMultipart(true);
-        assertTrue(req.isExpectMultipart());
-
-        req.uploadHandler(upload -> {
-
-          serverConn.set(req.connection());
-          checkClose.run();
-
-          Buffer tot = Buffer.buffer();
-          assertEquals("file", upload.name());
-          assertEquals(expectedFilename, upload.filename());
-          assertEquals("image/gif", upload.contentType());
-          String uploadedFileName;
-          if (!streamToDisk) {
-            upload.handler(tot::appendBuffer);
-            upload.exceptionHandler(err -> {
-              assertTrue(abortClient);
-              complete();
-            });
-            upload.endHandler(v -> {
-              assertFalse(abortClient);
-              assertEquals(content, tot);
-              assertTrue(upload.isSizeAvailable());
-              assertEquals(content.length(), upload.size());
-              assertNull(upload.file());
-              complete();
-            });
-          } else {
-            uploadedFileName = new File(testDir, UUID.randomUUID().toString()).getPath();
-            upload.streamToFileSystem(uploadedFileName, ar -> {
-              if (ar.succeeded()) {
-                Buffer uploaded = vertx.fileSystem().readFileBlocking(uploadedFileName);
-                assertEquals(content.length(), uploaded.length());
-                assertEquals(content, uploaded);
-                AsyncFile file = upload.file();
-                assertNotNull(file);
-                try {
-                  file.flush();
-                  fail("Was expecting uploaded file to be closed");
-                } catch (IllegalStateException ignore) {
-                  // File has been closed
-                }
-              } else {
-
-              }
-              complete();
-            });
-          }
-        });
-        req.endHandler(v -> {
-          MultiMap attrs = req.formAttributes();
-          attributeCount.set(attrs.size());
-          req.response().end();
-        });
-      }
-    });
-
-    server.listen(testAddress, onSuccess(s -> {
-      client.request(new RequestOptions(requestOptions)
-        .setMethod(HttpMethod.POST)
-        .setURI("/form"))
-        .onComplete(onSuccess(req -> {
-          String boundary = "dLV9Wyq26L_-JQxk6ferf-RT153LhOO";
-          String epi = "\r\n" +
-            "--" + boundary + "--\r\n";
-          String pro = "--" + boundary + "\r\n" +
-            "Content-Disposition: form-data; name=\"file\"" + (filename == null ? "" : "; filename=\"" + filename + "\"" ) + (extFilename == null ? "" : "; filename*=\"UTF-8''" + extFilename) + "\"\r\n" +
-            "Content-Type: image/gif\r\n" +
-            "\r\n";
-          req.headers().set("content-length", "" + (pro + contentStr + epi).length());
-          req.headers().set("content-type", "multipart/form-data; boundary=" + boundary);
-          if (abortClient) {
-            req.write(pro + contentStr.substring(0, contentStr.length() / 2), onSuccess(v -> {
-              clientConn.set(req.connection());
-              checkClose.run();
-            }));
-          } else {
-            req.end(pro + contentStr + epi);
-          }
-          req.response(ar -> {
-          assertEquals(ar.failed(), abortClient);
-          if (ar.succeeded()) {
-            HttpClientResponse resp = ar.result();
-            // assert the response
-            assertEquals(200, resp.statusCode());
-            resp.bodyHandler(body -> {
-              assertEquals(0, body.length());
-            });
-            assertEquals(0, attributeCount.get());
-          }
-          complete();
-        });
-      }));
-    }));
-    await();
-  }
-
-  @Test
-  public void testFormUploadAttributes() throws Exception {
-    AtomicInteger attributeCount = new AtomicInteger();
-    server.requestHandler(req -> {
-      if (req.method() == HttpMethod.POST) {
-        assertEquals(req.path(), "/form");
-        req.response().setChunked(true);
-        req.setExpectMultipart(true);
-        req.uploadHandler(upload -> upload.handler(buffer -> {
-          fail("Should get here");
-        }));
-        req.endHandler(v -> {
-          MultiMap attrs = req.formAttributes();
-          attributeCount.set(attrs.size());
-          assertEquals("vert x", attrs.get("framework"));
-          assertEquals("vert x", req.getFormAttribute("framework"));
-          assertEquals("jvm", attrs.get("runson"));
-          assertEquals("jvm", req.getFormAttribute("runson"));
-          req.response().end();
-        });
-      }
-    });
-
-    Buffer buffer = Buffer.buffer();
-    // Make sure we have one param that needs url encoding
-    buffer.appendString("framework=" + URLEncoder.encode("vert x", "UTF-8") + "&runson=jvm", "UTF-8");
-    server.listen(testAddress, onSuccess(s -> {
-      client.request(new RequestOptions(requestOptions)
-        .setMethod(HttpMethod.POST)
-        .setURI("/form"), onSuccess(req -> {
-          req
-            .putHeader("content-length", String.valueOf(buffer.length()))
-            .putHeader("content-type", "application/x-www-form-urlencoded")
-            .send(buffer, onSuccess(resp -> {
-            // assert the response
-            assertEquals(200, resp.statusCode());
-            resp.bodyHandler(body -> {
-              assertEquals(0, body.length());
-            });
-            assertEquals(2, attributeCount.get());
-            testComplete();
-          }));
-        }));
-    }));
-
-    await();
-  }
-
-  @Test
-  public void testFormUploadAttributes2() {
-    AtomicInteger attributeCount = new AtomicInteger();
-    server.requestHandler(req -> {
-      if (req.method() == HttpMethod.POST) {
-        assertEquals(req.path(), "/form");
-        req.setExpectMultipart(true);
-        req.uploadHandler(event -> event.handler(buffer -> {
-          fail("Should not get here");
-        }));
-        req.endHandler(v -> {
-          MultiMap attrs = req.formAttributes();
-          attributeCount.set(attrs.size());
-          assertEquals("junit-testUserAlias", attrs.get("origin"));
-          assertEquals("admin@foo.bar", attrs.get("login"));
-          assertEquals("admin", attrs.get("pass word"));
-          req.response().end();
-        });
-      }
-    });
-
-    server.listen(testAddress, onSuccess(s -> {
-      Buffer buffer = Buffer.buffer();
-      buffer.appendString("origin=junit-testUserAlias&login=admin%40foo.bar&pass+word=admin");
-      client.request(new RequestOptions(requestOptions)
-        .setMethod(HttpMethod.POST)
-        .setURI("/form")).onComplete(onSuccess(req -> {
-        req.putHeader("content-length", String.valueOf(buffer.length()))
-          .putHeader("content-type", "application/x-www-form-urlencoded")
-          .response(onSuccess(resp -> {
-            // assert the response
-            assertEquals(200, resp.statusCode());
-            resp.bodyHandler(body -> {
-              assertEquals(0, body.length());
-            });
-            assertEquals(3, attributeCount.get());
-            testComplete();
-          })).end(buffer);
-      }));
-    }));
-
-    await();
-  }
-
-  @Test
   public void testHostHeaderOverridePossible() {
     server.requestHandler(req -> {
       assertEquals("localhost:4444", req.host());
